Final Versions of my Part
These are the final versions of my GeneratedData function (to generate
model data), periFreq (to calculate the frequency of counts),
peristimGraph (to graph periFreq), and GDRecreation (to calculate J
values, actually compute the peristim stuff, and find p-values).
probFind was written by Garrett to calculate the probability of firing
for each direction for each neuron.